Improve PSK31 demodulation
The previous implementation was thrown together for expediency. This
revamp includes a much more effective filter, and feedback loop
bandwidths tuned from some testing of on-air signals (though it wasn't
very rigorous).

The API is rearranged to make room for planned work which includes an
incoherent detector, and a Viterbi decoder.